Add files via upload
Atliqo launched its 5G plans in May 2022, a significant move in the Indian telecom market. The objective of this project was to analyze Atliqo’s market share from January to September 2022 and determine whether their market presence was increasing, stable, or declining.

The analysis focused on comparing the company’s key performance indicators (KPIs) before and after the 5G launch, with a specific focus on revenue, active users, and market share.